+++ Tests/Fixtures/ClassRule/C.php
@@ -9,10 +9,10 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\ObjectMapper\Tests\Fixtures\MultipleTargetProperty;
+namespace Symfony\Component\ObjectMapper\Tests\Fixtures\ClassRule;
 
-class B
+class C
 {
-    public string $foo;
-    public string $otherFoo;
+    public string $foo = 'donotmap';
+    public string $bar = 'TESTSOURCED';
 }